Create BMS with remote monitoring
I've made this code for designing a BMS for lithium ion batteries, this code is capable of SOC, SOH measurement and you can monitor the health of your battery pack via a ThingSpeak server. The data collected on the server can be used for performance prediction, end of life prediction etc.